Update oc_leash.lsl
Utilizing llCastRay to identify line-of-sight between collar wearer and leash holder. With this  modification, re-leashing will only occur when there is line-of-sight between the two people, thus ensuring that if, for example, the leasher logs out and back in again on a higher floor of a building the leashee won't be dragged upwards and suspended from the ceiling by their neck. 

The same method is used during leashing if the leashee gets stuck. If for example the leash holder walks through a door and the door closes behind them before the leashee is able to get through, instead of the leashee being dragged into the door and being unable to move and possibly unable to open the door because they can't move far enough to give it space to open, or being stuck behind a wall, the leash script will now teleport them close to the leasher to catch up with them if there is a valid place to teleport them to close to the leasher (i.e teleport routing is not on for the parcel and there is open space near the leasher). If teleport is not possible the leash will remain visible but will no longer pull the leashee until they are back in line of site.

Other minor changes include a touch of streamlining, commenting out a function used only for debugging, and fixing the menu prompt so that it can now correctly report the leash holder when they are not in the sim.