PBR fix  for hide/show, move function out of themes
Allows Hide/Show to work with PBR materials. This uses a function that's bugged and will not properly revert alpha values correctly (bug reported to LL) so currently it toggles alpha values between 0 and 1. This is how the current hide/show works with BP textures anyway so no big issue, but in future when the llSetGLTFOverrides function is fixed, it will be possible to preserve alpha values for better functionality.

The hide/show has been moved to the states script and excised from the oc_themes script. This means that for collars that do not use themes, it's no longer necessary to have that script just to handle show/hide functionality.