Create oc_leashaddon
Replaces collar leashpoint

Collar still handles leashing, unleashing,  physics and leash settings. The addon leashpoint only redirects the particle chain from collar to the addon prim. 

Menu with Active(on/off), Lock(on/off) and AutoHide(on/of). The leashpoint lock is released once the leashpoint loses contact with the collar, this can take 5 minutes.

To use it simply drop the script into something like a septum ring and then wear it while the collar is used. The script will then connect as an addon for the collar. For detach functions, and autohide to work the script should be added to the root prim. If the intended leash prim is not the root, simply rename to “leashpoint”.